refactor: extract follow-up actions from AgentWorkspace
